Implement metadata reading
Both file and EXIF metadata seem to be successfully read.
Writing the output JSON is not yet implemented, and therefore
the integration tests are still failing.
Tweaked the FileMetadataOfInterest structure to reflect that some
platforms won't provide created and modified times; they are optional.
Updated the integration tests accordingly and fixed a couple of typos
and missed logging on errors.