Basic JSON writing of metadata works
Correct values are getting written to the output JSON.
The format is not yet correct because the individual members
of "MetadataOfInterest" are getting explicitly tagged and not
flattened.
Integration tests are additionally failing because they need the
created/modified dates to be set.
Otherwise seems to be OK, will fix the issues in next commit.