Integration tests pass
Add custom deserializer for Orientation so that the JSON
metadata can be successfully read by the integration tests.
(in hindsight, using a custom enum for Orientation may not have
been worth the extra trouble and code)

Also add some helpers in the tests to get created and modified
times for comparison against the output of rusimeta.

Integration tests now pass.